test: optimize wallet import tests (#204)
