+++ src/domains/wallet/pages/ImportWallet/ImportWallet.test.tsx
@@ -1,34 +1,29 @@
 /* eslint-disable @typescript-eslint/await-thenable */
 /* eslint-disable @typescript-eslint/require-await */
-import { Contracts, Wallet } from "@ardenthq/sdk-profiles";
+import { Contracts } from "@ardenthq/sdk-profiles";
 import userEvent from "@testing-library/user-event";
 import { createHashHistory } from "history";
 import nock from "nock";
 import React from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import { Route } from "react-router-dom";
-import { ToastContainer } from "react-toastify";
 
 import { ImportWallet } from "./ImportWallet";
 import { MethodStep } from "./MethodStep";
 import { SuccessStep } from "./SuccessStep";
 import { EnvironmentProvider, LedgerProvider } from "@/app/contexts";
 import { translations as commonTranslations } from "@/app/i18n/common/i18n";
-import { toasts } from "@/app/services";
 import { NetworkStep } from "@/domains/wallet/components/NetworkStep";
 import { OptionsValue } from "@/domains/wallet/hooks/use-import-options";
-import { translations as walletTranslations } from "@/domains/wallet/i18n";
 import { assertNetwork } from "@/utils/assertions";
 import {
 	env,
 	getDefaultProfileId,
-	MNEMONICS,
 	render,
 	renderResponsive,
 	renderWithForm,
 	screen,
 	waitFor,
-	within,
 	mockNanoXTransport,
 	mockProfileWithPublicAndTestNetworks,
 	mockProfileWithOnlyPublicNetworks,
@@ -37,19 +32,13 @@ import {
 let profile: Contracts.IProfile;
 const fixtureProfileId = getDefaultProfileId();
 
-const identityAddress = "DC8ghUdhS8w8d11K8cFQ37YsLBFhL3Dq2P";
 const mnemonic = "buddy year cost vendor honey tonight viable nut female alarm duck symptom";
 const randomAddress = "D61mfSggzbvQgTUe6JhYKH2doHaqJ3Dyib";
-const randomPublicKey = "034151a3ec46b5670a682b0a63394f863587d1bc97483b1b6c70eb58e7f0aed192";
-const randomPublicKeyInvalid = "a34151a3ec46b5670a682b0a63394f863587d1bc97483b1b6c70eb58e7f0aed192";
 
 const route = `/profiles/${fixtureProfileId}/wallets/import`;
 const routeLedger = `/profiles/${fixtureProfileId}/wallets/import/ledger`;
 const history = createHashHistory();
 
-jest.setTimeout(30_000);
-
-const enableEncryptionToggle = () => userEvent.click(screen.getByTestId("ImportWallet__encryption-toggle"));
 const continueButton = () => screen.getByTestId("ImportWallet__continue-button");
 const backButton = () => screen.getByTestId("ImportWallet__back-button");
 const mnemonicInput = () => screen.getByTestId("ImportWallet__mnemonic-input");
@@ -57,14 +46,10 @@ const addressInput = () => screen.findByTestId("ImportWallet__address-input");
 const finishButton = () => screen.getByTestId("ImportWallet__finish-button");
 const successStep = () => screen.getByTestId("ImportWallet__success-step");
 const methodStep = () => screen.getByTestId("ImportWallet__method-step");
-const publicKeyInput = () => screen.getByTestId("ImportWallet__publicKey-input");
 const privateKeyInput = () => screen.getByTestId("ImportWallet__privatekey-input");
 const wifInput = () => screen.getByTestId("ImportWallet__wif-input");
 const encryptedWifInput = () => screen.getByTestId("ImportWallet__encryptedWif-input");
 
-const secretInputID = "ImportWallet__secret-input";
-const errorText = "data-errortext";
-const password = "S3cUrePa$sword";
 const testNetwork = "ark.devnet";
 const ARKDevnet = "ARK Devnet";
 
@@ -350,698 +335,6 @@ describe("ImportWallet", () => {
 		});
 	});
 
-	it("should import by mnemonic", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.keyboard("{enter}");
-
-		await waitFor(() => expect(() => mnemonicInput()).not.toThrow());
-
-		expect(mnemonicInput()).toBeInTheDocument();
-
-		userEvent.paste(mnemonicInput(), mnemonic);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.keyboard("{enter}");
-
-		await waitFor(() => {
-			expect(successStep()).toBeInTheDocument();
-		});
-
-		userEvent.click(screen.getByTestId("ImportWallet__edit-alias"));
-
-		await expect(screen.findByTestId("Modal__inner")).resolves.toBeVisible();
-
-		userEvent.paste(screen.getByTestId("UpdateWalletName__input"), "test alias");
-
-		await waitFor(() => expect(screen.getByTestId("UpdateWalletName__submit")).toBeEnabled());
-
-		userEvent.keyboard("{enter}");
-		userEvent.click(screen.getByTestId("UpdateWalletName__submit"));
-
-		await waitFor(() => expect(screen.queryByTestId("Modal__inner")).not.toBeInTheDocument());
-
-		userEvent.click(finishButton());
-
-		await waitFor(() => {
-			expect(profile.wallets().findByAddressWithNetwork(identityAddress, testNetwork)).toBeInstanceOf(Wallet);
-		});
-	});
-
-	it("should import by mnemonic and use encryption password", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		expect(mnemonicInput()).toBeInTheDocument();
-
-		userEvent.paste(mnemonicInput(), MNEMONICS[3]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		userEvent.paste(screen.getByTestId("PasswordValidation__encryptionPassword"), password);
-		userEvent.paste(screen.getByTestId("PasswordValidation__confirmEncryptionPassword"), password);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(
-			() => {
-				expect(successStep()).toBeInTheDocument();
-			},
-			{ timeout: 15_000 },
-		);
-	});
-
-	it("should disable the encryption option when selecting a methods without encryption", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		expect(mnemonicInput()).toBeInTheDocument();
-
-		userEvent.paste(mnemonicInput(), MNEMONICS[0]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		expect(screen.getByTestId("ImportWallet__encryption-toggle")).not.toBeChecked();
-
-		enableEncryptionToggle();
-
-		expect(screen.getByTestId("ImportWallet__encryption-toggle")).toBeChecked();
-
-		// Select address that doesnt accept encryption
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.ADDRESS)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.ADDRESS));
-
-		await expect(addressInput()).resolves.toBeVisible();
-
-		expect(screen.getByTestId("ImportWallet__encryption-toggle")).not.toBeChecked();
-	});
-
-	it("should import by mnemonic with second signature and use password to encrypt both", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		expect(mnemonicInput()).toBeInTheDocument();
-
-		userEvent.paste(mnemonicInput(), MNEMONICS[0]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		userEvent.paste(screen.getByTestId("EncryptPassword__second-mnemonic"), MNEMONICS[5]);
-
-		userEvent.paste(screen.getByTestId("PasswordValidation__encryptionPassword"), password);
-		userEvent.paste(screen.getByTestId("PasswordValidation__confirmEncryptionPassword"), password);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(
-			() => {
-				expect(successStep()).toBeInTheDocument();
-			},
-			{ timeout: 15_000 },
-		);
-	});
-
-	it("should show an error message for invalid second mnemonic", async () => {
-		const walletId = profile
-			.wallets()
-			.findByAddressWithNetwork("DABCrsfEqhtdzmBrE2AU5NNmdUFCGXKEkr", testNetwork)
-			?.id();
-		if (walletId) {
-			profile.wallets().forget(walletId);
-		}
-
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-				withProviders: true,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		userEvent.paste(mnemonicInput(), MNEMONICS[0]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		userEvent.paste(
-			screen.getByTestId("EncryptPassword__second-mnemonic"),
-			"invalid second mnemonic fjdkfjdkjfkdjf",
-		);
-
-		await waitFor(() => {
-			expect(screen.getAllByTestId("Input__error")[0]).toHaveAttribute(
-				errorText,
-				walletTranslations.PAGE_IMPORT_WALLET.VALIDATION.INVALID_MNEMONIC,
-			);
-		});
-	});
-
-	it("should import by address", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.ADDRESS)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.ADDRESS));
-
-		await expect(addressInput()).resolves.toBeVisible();
-
-		userEvent.paste(await addressInput(), randomAddress);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(successStep()).toBeInTheDocument();
-		});
-
-		userEvent.click(finishButton());
-
-		await waitFor(() => {
-			expect(profile.wallets().findByAddressWithNetwork(randomAddress, testNetwork)).toBeInstanceOf(Wallet);
-		});
-	});
-
-	it("should import by public key", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.PUBLIC_KEY)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.PUBLIC_KEY));
-
-		await expect(screen.findByTestId("ImportWallet__publicKey-input")).resolves.toBeVisible();
-
-		userEvent.paste(publicKeyInput(), randomPublicKey);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(successStep()).toBeInTheDocument();
-		});
-
-		userEvent.click(finishButton());
-
-		await waitFor(() => {
-			expect(profile.wallets().findByAddressWithNetwork(randomAddress, testNetwork)).toBeInstanceOf(Wallet);
-		});
-	});
-
-	it("should validate public key doesnt exist", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.PUBLIC_KEY)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.PUBLIC_KEY));
-
-		await expect(screen.findByTestId("ImportWallet__publicKey-input")).resolves.toBeVisible();
-
-		const findAdressSpy = jest.spyOn(profile.wallets(), "findByAddressWithNetwork").mockReturnValue({} as any);
-
-		userEvent.paste(publicKeyInput(), randomPublicKey);
-
-		await waitFor(() => expect(continueButton()).toBeDisabled());
-
-		findAdressSpy.mockRestore();
-	});
-
-	it("should not allow importing from an invalid public key", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.PUBLIC_KEY)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.PUBLIC_KEY));
-
-		await expect(screen.findByTestId("ImportWallet__publicKey-input")).resolves.toBeVisible();
-
-		userEvent.paste(publicKeyInput(), randomPublicKeyInvalid);
-
-		await waitFor(() => expect(continueButton()).toBeDisabled());
-	});
-
-	it("should import by secret", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		const countBefore = profile.wallets().count();
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.SECRET)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.SECRET));
-
-		await expect(screen.findByTestId(secretInputID)).resolves.toBeVisible();
-
-		userEvent.paste(screen.getByTestId(secretInputID), "secret.111");
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(successStep()).toBeInTheDocument();
-		});
-
-		userEvent.click(finishButton());
-
-		await waitFor(() => expect(profile.wallets().count()).toBe(countBefore + 1));
-	});
-
-	it("should import by secret with encryption", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.SECRET)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.SECRET));
-
-		await expect(screen.findByTestId(secretInputID)).resolves.toBeVisible();
-
-		userEvent.paste(screen.getByTestId(secretInputID), "secret.222");
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		userEvent.paste(screen.getByTestId("PasswordValidation__encryptionPassword"), password);
-		userEvent.paste(screen.getByTestId("PasswordValidation__confirmEncryptionPassword"), password);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(
-			() => {
-				expect(successStep()).toBeInTheDocument();
-			},
-			{ timeout: 15_000 },
-		);
-	});
-
-	it("should import by secret with second signature and use password to encrypt both", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.SECRET)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.SECRET));
-
-		expect(methodStep()).toBeInTheDocument();
-
-		const passphraseInput = screen.getByTestId(secretInputID);
-
-		expect(passphraseInput).toBeInTheDocument();
-
-		userEvent.paste(passphraseInput, "abc");
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		userEvent.paste(screen.getByTestId("EncryptPassword__second-secret"), "abc");
-
-		userEvent.paste(screen.getByTestId("PasswordValidation__encryptionPassword"), password);
-		userEvent.paste(screen.getByTestId("PasswordValidation__confirmEncryptionPassword"), password);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(
-			() => {
-				expect(successStep()).toBeInTheDocument();
-			},
-			{ timeout: 15_000 },
-		);
-	});
-
-	it("forgets the imported wallet if back from encrypted password step", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.SECRET)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.SECRET));
-
-		expect(methodStep()).toBeInTheDocument();
-
-		const passphraseInput = screen.getByTestId(secretInputID);
-
-		expect(passphraseInput).toBeInTheDocument();
-
-		userEvent.paste(passphraseInput, "abcd");
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		const profileForgetWalletSpy = jest.spyOn(profile.wallets(), "forget").mockImplementation(() => {});
-
-		userEvent.click(backButton());
-
-		expect(profileForgetWalletSpy).toHaveBeenCalledWith(expect.any(String));
-
-		profileForgetWalletSpy.mockRestore();
-	});
-
-	it("should show an error message for invalid second secret", async () => {
-		const walletId = profile
-			.wallets()
-			.findByAddressWithNetwork("DNTwQTSp999ezQ425utBsWetcmzDuCn2pN", testNetwork)
-			?.id();
-		if (walletId) {
-			profile.wallets().forget(walletId);
-		}
-
-		const wallet = await profile.walletFactory().fromSecret({
-			coin: "ARK",
-			network: "ark.devnet",
-			secret: "abc",
-		});
-
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.SECRET)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.SECRET));
-
-		expect(methodStep()).toBeInTheDocument();
-
-		const passphraseInput = screen.getByTestId(secretInputID);
-
-		expect(passphraseInput).toBeInTheDocument();
-
-		userEvent.paste(passphraseInput, "abc");
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-
-		enableEncryptionToggle();
-
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("EncryptPassword")).toBeInTheDocument();
-		});
-
-		const fromSecretMock = jest.spyOn(wallet.coin().address(), "fromSecret").mockImplementationOnce(() => {
-			throw new Error("test");
-		});
-
-		userEvent.paste(screen.getByTestId("EncryptPassword__second-secret"), "invalid second secret");
-
-		await waitFor(() => {
-			expect(screen.getAllByTestId("Input__error")[0]).toHaveAttribute(
-				errorText,
-				walletTranslations.PAGE_IMPORT_WALLET.VALIDATION.INVALID_SECRET,
-			);
-		});
-
-		fromSecretMock.mockRestore();
-	});
-
 	it("should get options depend on the network", async () => {
 		render(
 			<Route path="/profiles/:profileId/wallets/import">
@@ -1073,136 +366,6 @@ describe("ImportWallet", () => {
 		await waitFor(() => expect(screen.queryByText(commonTranslations.PRIVATE_KEY)).not.toBeInTheDocument());
 		await waitFor(() => expect(screen.queryByText(commonTranslations.WIF)).not.toBeInTheDocument());
 		await waitFor(() => expect(screen.queryByText(commonTranslations.ENCRYPTED_WIF)).not.toBeInTheDocument());
-	});
-
-	it("should show an error message for duplicate address when importing by mnemonic", async () => {
-		const generated = await profile.walletFactory().generate({
-			coin: "ARK",
-			network: testNetwork,
-		});
-
-		profile.wallets().push(generated.wallet);
-
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		expect(mnemonicInput()).toBeInTheDocument();
-
-		userEvent.paste(mnemonicInput(), generated.mnemonic);
-
-		await waitFor(() => {
-			expect(screen.getByTestId("Input__error")).toHaveAttribute(
-				errorText,
-				commonTranslations.INPUT_ADDRESS.VALIDATION.ADDRESS_ALREADY_EXISTS.replace(
-					"{{address}}",
-					generated.wallet.address(),
-				),
-			);
-		});
-
-		expect(continueButton()).toBeDisabled();
-	});
-
-	it("should show an error message for duplicate address when importing by address", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.ADDRESS)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.ADDRESS));
-
-		await expect(addressInput()).resolves.toBeVisible();
-
-		userEvent.paste(await addressInput(), profile.wallets().first().address());
-
-		await waitFor(() => {
-			expect(screen.getByTestId("Input__error")).toHaveAttribute(
-				errorText,
-				commonTranslations.INPUT_ADDRESS.VALIDATION.ADDRESS_ALREADY_EXISTS.replace(
-					"{{address}}",
-					profile.wallets().first().address(),
-				),
-			);
-		});
-
-		expect(continueButton()).toBeDisabled();
-	});
-
-	it("should show an error message for invalid address", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.ADDRESS)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.ADDRESS));
-
-		await expect(addressInput()).resolves.toBeVisible();
-
-		userEvent.paste(await addressInput(), "123");
-
-		await waitFor(() => {
-			expect(screen.getByTestId("Input__error")).toHaveAttribute(
-				errorText,
-				commonTranslations.INPUT_ADDRESS.VALIDATION.NOT_VALID,
-			);
-		});
-
-		expect(continueButton()).toBeDisabled();
 	});
 
 	it("should render as ledger import", async () => {
@@ -1284,124 +447,6 @@ describe("ImportWallet", () => {
 		});
 
 		historySpy.mockRestore();
-	});
-
-	it("should show an error message for duplicate name", async () => {
-		const emptyProfile = await env.profiles().create("duplicate wallet name profile");
-
-		await env.profiles().restore(emptyProfile);
-		await emptyProfile.sync();
-
-		const randomNewAddress = "D6pPxYLwwCptuhVRvLQQYXEQiQMB5x6iY3";
-
-		const wallet = await emptyProfile.walletFactory().fromMnemonicWithBIP39({
-			coin: "ARK",
-			mnemonic: MNEMONICS[1],
-			network: testNetwork,
-		});
-
-		wallet.settings().set(Contracts.WalletSetting.Alias, "My wallet");
-
-		profile.wallets().push(wallet);
-
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<ImportWallet />
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => expect(() => methodStep()).not.toThrow());
-
-		expect(methodStep()).toBeInTheDocument();
-
-		userEvent.click(screen.getByTestId("SelectDropdown__caret"));
-
-		await expect(screen.findByText(commonTranslations.ADDRESS)).resolves.toBeVisible();
-
-		userEvent.click(screen.getByText(commonTranslations.ADDRESS));
-
-		await expect(addressInput()).resolves.toBeVisible();
-
-		userEvent.paste(await addressInput(), randomNewAddress);
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await waitFor(() => {
-			expect(successStep()).toBeInTheDocument();
-		});
-
-		const alias = "My Wallet";
-
-		userEvent.click(screen.getByTestId("ImportWallet__edit-alias"));
-
-		await expect(screen.findByTestId("UpdateWalletName__input")).resolves.toBeVisible();
-
-		// Test cancel button
-		userEvent.click(screen.getByTestId("UpdateWalletName__cancel"));
-
-		expect(screen.queryByTestId("UpdateWalletName__input")).toBeNull();
-
-		// Try to edit name again
-		userEvent.click(screen.getByTestId("ImportWallet__edit-alias"));
-
-		userEvent.clear(screen.getByTestId("UpdateWalletName__input"));
-		userEvent.paste(screen.getByTestId("UpdateWalletName__input"), alias);
-
-		await waitFor(() => {
-			expect(screen.getByTestId("Input__error")).toHaveAttribute(
-				errorText,
-				walletTranslations.VALIDATION.ALIAS_ASSIGNED.replace("{{alias}}", alias),
-			);
-		});
-
-		expect(screen.getByTestId("UpdateWalletName__submit")).toBeDisabled();
-	});
-
-	it("should show warning sync error toast in network step and retry sync", async () => {
-		render(
-			<Route path="/profiles/:profileId/wallets/import">
-				<>
-					<ToastContainer closeOnClick={false} newestOnTop />
-					<ImportWallet />
-				</>
-			</Route>,
-			{
-				route: route,
-			},
-		);
-
-		await expect(screen.findByTestId("NetworkStep")).resolves.toBeVisible();
-
-		userEvent.click(screen.getAllByTestId("NetworkOption")[1]);
-
-		const coin = profile.coins().get("ARK", testNetwork);
-		const coinMock = jest.spyOn(coin, "__construct").mockImplementationOnce(() => {
-			throw new Error("test");
-		});
-
-		await waitFor(() => expect(continueButton()).toBeEnabled());
-		userEvent.click(continueButton());
-
-		await expect(screen.findByTestId("SyncErrorMessage__retry")).resolves.toBeVisible();
-
-		const toastDismissMock = jest.spyOn(toasts, "dismiss").mockResolvedValue(undefined);
-		userEvent.click(within(screen.getByTestId("SyncErrorMessage__retry")).getByRole("link"));
-
-		await expect(screen.findByTestId("SyncErrorMessage__retry")).resolves.toBeVisible();
-
-		coinMock.mockRestore();
-		toastDismissMock.mockRestore();
 	});
 
 	describe("import with private key", () => {
